// frontend/src/app/model/komentar.ts
export class Komentar {
    idKnjige: number;
    tekst: string;
    username: string;
    datum: Date;
    dan: number;
    mesec: number;
    godina: number;
    sati: number;
    minuti: number;
    azurirano: boolean;
    datumAzuriranja: Date;
    danA: number;
    mesecA: number;
    godinaA: number;
    satiA: number;
    minutiA: number;

    godinaString: string;
    mesecString: string;
    danString: string;
    satiString: string;
    minutiString: string;

    godinaAString: string;
    mesecAString: string;
    danAString: string;
    satiAString: string;
    minutiAString: string;


}